feat: add ConfigContext, encapsulate feedback form logic in the plugin
