+++ typescript/ui/src/contexts/ConfigContext/ConfigContext.ts
@@ -1,0 +1,10 @@
+import { ConfigResponse } from "ragbits-api-client";
+import { createContext } from "react";
+
+export interface IConfigContext {
+  config: ConfigResponse;
+}
+
+export const ConfigContext = createContext<IConfigContext | undefined>(
+  undefined,
+);
