+++ typescript/ui/src/contexts/ConfigContext/ConfigContextProvider.tsx
@@ -1,0 +1,66 @@
+import { PropsWithChildren, useMemo } from "react";
+import { ConfigContext } from "./ConfigContext";
+import { useRagbitsCall } from "ragbits-api-client-react";
+import { useThemeContext } from "../ThemeContext/useThemeContext";
+import { CircularProgress, cn } from "@heroui/react";
+
+export function ConfigContextProvider({ children }: PropsWithChildren) {
+  const { theme } = useThemeContext();
+  const { call: fetchConfig, ...config } = useRagbitsCall("/api/config");
+
+  const value = useMemo(() => {
+    if (!config.data) {
+      return null;
+    }
+
+    return {
+      config: config.data,
+    };
+  }, [config.data]);
+
+  // Load config on mount
+  if (!config.data && !config.error && !config.isLoading) {
+    fetchConfig();
+  }
+
+  // TODO: Consider adding minimal timeout for config to not flash users with this screen
+  if (config.isLoading) {
+    return (
+      <div
+        className={cn(
+          "flex h-screen w-screen items-start justify-center bg-background",
+          theme,
+        )}
+      >
+        <div className="m-auto flex flex-col items-center gap-4 text-default-900">
+          <CircularProgress size="lg" />
+          <p>Initializing...</p>
+        </div>
+      </div>
+    );
+  }
+
+  if (config.error || !value) {
+    return (
+      <div
+        className={cn(
+          "flex h-screen w-screen items-start justify-center bg-background",
+          theme,
+        )}
+      >
+        <div className="m-auto flex flex-col items-center gap-4 text-default-900">
+          <p className="text-large">
+            Something went wrong during chat initialization.
+          </p>
+          <p className="text-small text-default-500">
+            Try refreshing the page.
+          </p>
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <ConfigContext.Provider value={value}>{children}</ConfigContext.Provider>
+  );
+}
